Refactor String html methods
Move raw methods to a separate base file, and split out tests.
